refactor ftable
