--- include/dsp/ftable/deck.h
@@ -1,87 +1,0 @@
-/**
- * @brief ftable decks for various utilities
- *  - sinesum (band limited / novel waveforms)
- *  - TODO: interpolation / spectral morph (based on serum2)
- *  - TODO: wrapers for other ftable module functionality (cheby, window)
- *
- */
-
-#ifndef DSP_DECK_H
-#define DSP_DECK_H
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-#include <dsp/fasttrig.h>
-#include <dsp/ftable/ftable.h>
-#include <dsp/ftable/sinesum.h>
-#include <dsp/maths.h>
-#include <dsp/rmap.h>
-#include <dsp/xfade.h>
-
-#include <stdint.h>
-
-/**
- * @brief ft_deck is fat pointer type that simply holds an array of ftable pointers.
- * Its lookup strategy is based on position. We borrow terminology from serum and other
- * wt synths and call these "frames".
- */
-typedef struct {
-    ftable** frames;
-    uint32_t frames_sz;  // nrows
-} ft_deck;
-
-/**
- * @brief init a ft_deck. We check that frames is at least 2 so that we can iterate
- * and determine if all frames in the deck are equal.
- */
-void ft_deck_init(ft_deck* self, ftable** frames, uint32_t frames_sz);
-
-/**
- * @brief check if all frames are of equal length. Usually this is the case when we're
- * using decks, but there could be impls that might want decks with frames of different
- * sizes.
- */
-bool ft_deck_frames_equal(ft_deck* self);
-
-/**
- * @brief fill a matrix from the underlying frames ptr. All frames must be equally
- * sized and the matrix should be initialized with the correct buffer
- */
-dsp_err ft_deck_matrix_fill(ft_deck* self, matrix* out);
-
-/**
- * @brief represents a low/high pair. The tables used for the actual crossfade.
- */
-typedef struct {
-    ftable* low;
-    ftable* high;
-} ft_frame_pair;
-
-/**
- * @brief return a frame_pair from a deck given pos.
- */
-static inline ft_frame_pair ft_deck_pos_lookup(ft_deck* self, float pos) {
-    pos = clamp(pos, 0.0f, 1.0f);         // ensure [0, 1]
-    pos = 0.5f - 0.5f * fast_hcosf(pos);  // smooth transition w/ half cos
-
-    float fidx = pos * (self->frames_sz - 1);  // range: [0, N-1]
-    uint32_t idx = (int) fidx;
-
-    if (idx >= self->frames_sz - 1) {
-        idx = self->frames_sz - 2;
-        fidx = (float) idx;
-    }
-
-    return (ft_frame_pair) {
-        .low = self->frames[idx],
-        .high = self->frames[idx + 1],
-    };
-}
-
-#ifdef __cplusplus
-}
-#endif
-
-#endif
